+++ src/compas_timber/fabrication/lap.py
@@ -710,7 +710,7 @@ class Lap(BTLxProcessing):
         assert self.slope is not None
 
         ref_surface = beam.side_as_surface(self.ref_side_index)
-        p_origin = ref_surface.point_at(self.start_x, self.start_y)
+        p_origin = ref_surface.point_at(self.start_x / ref_surface.xsize, self.start_y / ref_surface.ysize)
         start_frame = Frame(p_origin, -ref_surface.frame.yaxis, ref_surface.frame.xaxis)
 
         # define angle rotation matrix